feat: adding srs APIs, refactoring kas APIs
